// src/pages/main-menu/wholesale-monitor/models/interfaces.ts
import { SubsEventList } from "../../data-reference/event/hooks/EventContext";

export interface WholesaleMonitorList {
  startNbr?: string;
  rowno: string;
  comments?: string;
  priority?: string;
  partyType: string;
  spId: number;
  extAttr: string;
  subsPlanId?: string;
  subsEventId: string;
  createdDate: string;
  reqDate: string;
  keyStaff: string;
  wholesaleId: string;
  endNbr?: string;
  partyCode: string;
  eventName: string;
  offerId?: string;
  state: string;
  wholesaleCode: string;
  commisionAmount?: string;
}

export interface BatchDetailList {
  rowno: string;
  orderNbr: string;
  simCardId: string;
  orderItemId: string;
  rowno1: string;
  completedDate: string;
  orderState: string;
  iccid: string;
  subsId: string;
  wholesaleInstId: string;
  wholesaleId: string;
  accNbr: string;
  stateDate: string;
  state: string;
  accNbrId: string;
  orderStateName: string;
  comments?: string;
}

export interface DomainList {
  tableName: string;
  columnName: string;
  value: string;
  lookupName: string;
  comments: string;
}

export interface WholesaleMasterData {
  subsEvent: SubsEventList[];
  queryState: DomainList[];
  orderState: OrderStateList[];
}

export interface OrgList {
  orgId: number;
  areaId: number;
  orgName: string;
  orgType: string;
  state: string;
  stateDate: string;
  orgCode: string;
  spId: number;
}

export interface StaffList {
  STAFF_ORG_ID: number;
  USER_CODE: string;
  EXIST: number;
  CREATED_DATE: string;
  STATE_DATE: string;
  STAFF_ID: number;
  STATE: string;
  USER_ID: number;
  USER_NAME: string;
  SP_ID?: number;
  STAFF_NAME: string;
}

// Generated by https://quicktype.io

export interface WholesaleQuery {
  wholesaleId?: number;
  wholesaleCode?: string;
  partyType?: string;
  state?: string;
  subsEventId?: number;
  subsEventIdList?: number[];
  spId?: number;
  custId?: number;
  custName?: string;
  staffId?: number;
  executionDateFrom?: string;
  executionDateTo?: string;
  createDateFrom?: string;
  createDateTo?: string;
  accNbrBegin?: string;
  accNbrEnd?: string;
  iccidBegin?: string;
  iccidEnd?: string;
}

export interface OrderStateList {
  comments: string;
  orderState: string;
  orderStateName: string;
}